Event Color Selector & Tag Addition
I made a function to randomly assign a color to a new event. It only
has 10 different colors to chose from, but thats a fairly decent amount
as of now.
I also added a list view which populates with all of the bulbs on the
network for the tag system. Still needs some work, but it should allow
a user to select the bulbs they wish to edit.